Platform Sensors Implementation
The platform class is now responsible for providing the environmental
sensor status.

The previous method was simply to use labels in sensors3.conf and
scrub the output of 'sensors' a little, but this caused occasional errors and
transient data states to be displayed to the user. It also kept us
from providing useful labels for which parts of the system the sensors
were attached to (chassis, psu, etc).

The switchlight vendor-config package now provides a sensors.py package
which provides basic parsing and processing of the output of 'sensors'.

This package is now used to produce the platform-specific descriptions
of the sensors data for the user.