PAN-1074
The SLREST package now contains a localconfig module.

The localconfig module defines the names of local options
that may be set which affect SLREST API behavior.

Local configuration options are documented in the localconfig module
and are backed by touchfiles in the localconfig directory.

With the current support you must touch files manually in the localconfig directory.

The localconfig module can be imported into the CLI to provide a user interface
to these settings as the remaining step for this PR.