arpa: only check source if ARP packet-in reason is set
See BVS-2966. ARP requests from uplink ports can go to the CPU because of the
arp_offload table. We shouldn't do a source check on these packets because the
ARP table is only populated with entries local to the vswitch. IVS will start
sending arp_offload packets with a different packet-in reason so that we can
distinguish between these cases.