arpa: separate time fields for stats and timeouts
There are situations like table entry modify and cyclical timeouts where we
want to change the time used as the base for timeouts without losing the
information of the real last active time.

The only functional change in this commit is that table entry modify won't
reset the idle time reported to the controller.