arpa: cyclical timeouts
See PAN-2013.

Previously, once an entry got into the idle_timeout state it would only leave
that state if it was hit by an ARP packet. It would not continue to send
requeries. This caused problems in dual-TOR scenarios when the switch that had
been sending ARP requeries and receiving the replies shut down, because the
remaining switch would not send any queries. The fix is to continue sending
requeries after the idle notification.